+++ tests/Database/TableTest.php
@@ -284,7 +284,7 @@ abstract class TableTest extends BaseTest
         );
 
         $this->assertSame(4, $table->count());
-        $this->assertSame(100, (int)$table->sum('value'));
+        $this->assertSame(100, (int) $table->sum('value'));
     }
 
     public function testAggregationMinByPass(): void
@@ -303,7 +303,7 @@ abstract class TableTest extends BaseTest
         );
 
         $this->assertSame(4, $table->count());
-        $this->assertSame(10, (int)$table->min('value'));
+        $this->assertSame(10, (int) $table->min('value'));
     }
 
     public function testAggregationMaxByPass(): void
@@ -322,7 +322,7 @@ abstract class TableTest extends BaseTest
         );
 
         $this->assertSame(4, $table->count());
-        $this->assertSame(40, (int)$table->max('value'));
+        $this->assertSame(40, (int) $table->max('value'));
     }
 
     public function testAggregationAvgByPass(): void
@@ -341,7 +341,7 @@ abstract class TableTest extends BaseTest
         );
 
         $this->assertSame(4, $table->count());
-        $this->assertSame(25, (int)$table->avg('value'));
+        $this->assertSame(25, (int) $table->avg('value'));
     }
 
     public function testAggregationAvgByPassFloat(): void
@@ -360,7 +360,7 @@ abstract class TableTest extends BaseTest
         );
 
         $this->assertSame(4, $table->count());
-        $this->assertSame(13.75, (float)$table->avg('value'));
+        $this->assertSame(13.75, (float) $table->avg('value'));
     }
 
     public function testDeleteWithWhere(): void